clean architecture - start of global refactoring
