+++ lib/data/repository.dart
@@ -1,0 +1,6 @@
+import '../domain/sight.dart';
+
+class Repository {
+
+}
+
