+++ lib/ui/screens/main_screen/3_0_visiting_screen.dart
@@ -1,6 +1,6 @@
 import "package:flutter/material.dart";
 
-import '../elements/headers.dart';
+import '../../widgets/headers.dart';
 
 import '3_1_tab_wished.dart';
 import '3_2_tab_seen.dart';
