+++ lib/ui/screens/main_screen/main_screen.dart
@@ -1,0 +1,72 @@
+import 'package:flutter/material.dart';
+
+import '1_sight_list_screen.dart';
+import '2_map_screen.dart';
+import '3_0_visiting_screen.dart';
+import '4_settings_screen.dart';
+
+class AllMainScreens extends StatefulWidget {
+  @override
+  _AllMainScreensState createState() => new _AllMainScreensState();
+}
+
+class _AllMainScreensState extends State<AllMainScreens> {
+  int _page = 0;
+  PageController _c;
+  @override
+  void initState() {
+    _c = new PageController(
+      initialPage: _page,
+    );
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      bottomNavigationBar: BottomNavigationBar(
+        currentIndex: _page,
+        showSelectedLabels: false,
+        showUnselectedLabels: false,
+        onTap: (index) {
+          this._c.animateToPage(index,
+              duration: const Duration(milliseconds: 500),
+              curve: Curves.easeInOut);
+        },
+        type: BottomNavigationBarType.fixed,
+        items: [
+          BottomNavigationBarItem(
+            icon: Icon(Icons.list),
+            label: 'Список мест',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.map),
+            label: 'Карта',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.favorite_border),
+            label: 'Избранное',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.settings),
+            label: 'Настройки',
+          ),
+        ],
+      ),
+      body: PageView(
+        controller: _c,
+        onPageChanged: (newPage) {
+          setState(() {
+            this._page = newPage;
+          });
+        },
+        children: [
+          Center(child: SightListScreen()),
+          Center(child: MapsScreen()),
+          Center(child: VisitingScreen()),
+          Center(child: SettingsScreen()),
+        ],
+      ),
+    );
+  }
+}
