+++ lib/ui/widgets/sight_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import '../../domain/sight.dart';
-import '../../common.dart';
-import '../../routes.dart';
+import '../my_enums.dart';
+import '../my_app/routes.dart';
 import 'image.dart';
 
 import 'sight_card_icons.dart';
